finished

+++ src/main/java/com/boostore/jpa/repository/AuthorRepository.java
@@ -3,5 +3,5 @@ package com.boostore.jpa.repository;
 import com.boostore.jpa.models.AuthorModel;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AuthorRepository extends JpaRepository<AuthorModel, Integer > {
+public interface AuthorRepository extends JpaRepository<AuthorModel, Long > {
 }
