+++ src/main/java/com/boostore/jpa/repository/PublisherRepository.java
@@ -4,5 +4,5 @@ import com.boostore.jpa.models.AuthorModel;
 import com.boostore.jpa.models.PublisherModel;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PublisherRepository extends JpaRepository<PublisherModel, Integer > {
+public interface PublisherRepository extends JpaRepository<PublisherModel, Long > {
 }
