+++ src/main/java/com/boostore/jpa/repository/ReviewRepository.java
@@ -4,5 +4,5 @@ import com.boostore.jpa.models.AuthorModel;
 import com.boostore.jpa.models.ReviewModel;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ReviewRepository extends JpaRepository<ReviewModel, Integer > {
+public interface ReviewRepository extends JpaRepository<ReviewModel, Long > {
 }
